+++ src/ingestion/index.ts
@@ -10,6 +10,7 @@ export { jcbMyJcbAdapter } from './adapters/jcbMyJcb'
 export { smbcVpassAdapter } from './adapters/smbcVpass'
 export { securitiesAssetSnapshotAdapter } from './adapters/securitiesAssetSnapshot'
 export { sbiSecuritiesTradeHistoryAdapter } from './adapters/sbiSecuritiesTradeHistory'
+export { rakutenSecuritiesDomesticTradeHistoryAdapter } from './adapters/rakutenSecuritiesDomesticTradeHistory'
 export { japaneseBrokerageTransactionsAdapter } from './adapters/japaneseBrokerageTransactions'
 export { moneyForwardAssetTrendAdapter } from './adapters/moneyForwardAssetTrend'
 export { moneyForwardHouseholdLedgerAdapter } from './adapters/moneyForwardHouseholdLedger'
@@ -24,6 +25,7 @@ import { jcbMyJcbAdapter } from './adapters/jcbMyJcb'
 import { smbcVpassAdapter } from './adapters/smbcVpass'
 import { securitiesAssetSnapshotAdapter } from './adapters/securitiesAssetSnapshot'
 import { sbiSecuritiesTradeHistoryAdapter } from './adapters/sbiSecuritiesTradeHistory'
+import { rakutenSecuritiesDomesticTradeHistoryAdapter } from './adapters/rakutenSecuritiesDomesticTradeHistory'
 import { japaneseBrokerageTransactionsAdapter } from './adapters/japaneseBrokerageTransactions'
 import { moneyForwardAssetTrendAdapter } from './adapters/moneyForwardAssetTrend'
 import { moneyForwardHouseholdLedgerAdapter } from './adapters/moneyForwardHouseholdLedger'
@@ -39,6 +41,7 @@ export const importAdapters = [
   smbcVpassAdapter,
   securitiesAssetSnapshotAdapter,
   sbiSecuritiesTradeHistoryAdapter,
+  rakutenSecuritiesDomesticTradeHistoryAdapter,
   japaneseBrokerageTransactionsAdapter,
   moneyForwardAssetTrendAdapter,
   moneyForwardHouseholdLedgerAdapter,